Move WS proxy processing to a task.
