+++ mles-client/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "mles-client"
-version = "1.1.1"
+version = "1.1.2"
 authors = ["jq-rs"]
 edition = "2018"
