+++ mles-client/src/main.rs
@@ -39,12 +39,7 @@ use std::{env, process};
 use std::io::{self, Read, Write};
 use std::net::{IpAddr, Ipv4Addr, SocketAddr, ToSocketAddrs};
 use std::thread;
-use std::io::{Error, ErrorKind};
-use futures::sync::mpsc::{UnboundedSender, UnboundedReceiver};
 use std::time::Duration;
-use std::collections::HashMap;
-use std::rc::Rc;
-use std::cell::RefCell;
 
 use bytes::BytesMut;
 use futures::{Sink, Future, Stream};
@@ -302,128 +297,3 @@ fn read_stdin(mut rx: mpsc::Sender<Vec<u8>>) {
         rx = rx.send(msg).wait().unwrap();
     }
 }
-
-struct ChannelDb {
-    cid: u32,
-    key: u64,
-    socket: UnboundedSender<Vec<u8>>,
-}
-
-pub fn process_mles_multiclient(raddr: SocketAddr, keyval: String, keyaddr: String,
-                                ws_tx: UnboundedSender<Vec<u8>>, mles_rx: UnboundedReceiver<Vec<u8>>) {
-    let mut runtime = Runtime::new().unwrap();
-    let mut cid: Option<u32> = None;
-    let mut key: Option<u64> = None;
-    let mut keys: Vec<String> = Vec::new();
-    let uc_hash: HashMap<String, ChannelDb> = HashMap::new();
-    let uc_db = Rc::new(RefCell::new(uc_hash));
-    
-    let mles_rx = mles_rx.for_each(move |buf| {
-        //if buf.is_empty() {
-        //    return Err(Error::new(ErrorKind::BrokenPipe, "broken pipe"));
-        //}
-        let decoded_message = Msg::decode(buf.as_slice());
-        let uid = decoded_message.get_uid().to_string();
-        let channel = decoded_message.get_channel().to_string();
-        let key = uid + &channel;
-        let mut uc_db_once = uc_db.borrow_mut();
-        //verify do we need new connection
-        if !uc_db_once.contains_key(&key) {
-            //spawn a new task with the actual connection,
-            //bundle channel with it,
-            //add to hash
-            //
-            let tcp = TcpStream::connect(&raddr);
-        }
-        //send the buf to the task
-        Ok(())
-    }).map_err(|e| { println!("Got error {:?}!", e); });
-
-    runtime.spawn(mles_rx);
-
-    match runtime.run() {
-        Ok(_) => {}
-        Err(err) => {
-            println!("Error: {}", err);
-        }
-    };
-}
-
-
-pub fn process_mles_client(raddr: SocketAddr, keyval: String, keyaddr: String,
-                           ws_tx: UnboundedSender<Vec<u8>>, mles_rx: UnboundedReceiver<Vec<u8>>) {
-    let mut runtime = Runtime::new().unwrap();
-    let tcp = TcpStream::connect(&raddr);
-    let mut cid: Option<u32> = None;
-    let mut key: Option<u64> = None;
-    let mut keys = Vec::new();
-
-    let client = tcp.and_then(move |stream| {
-        let _val = stream.set_nodelay(true)
-                         .map_err(|_| panic!("Cannot set to no delay"));
-        let _val = stream.set_keepalive(Some(Duration::new(KEEPALIVE, 0)))
-                         .map_err(|_| panic!("Cannot set keepalive"));
-        let laddr = match stream.local_addr() {
-            Ok(laddr) => laddr,
-            Err(_) => {
-                let addr = "0.0.0.0:0";
-                addr.parse::<SocketAddr>().unwrap()
-            }
-        };
-        if  !keyval.is_empty() {
-            keys.push(keyval);
-        } else {
-            keys.push(MsgHdr::addr2str(&laddr));
-            if !keyaddr.is_empty() {
-                keys.push(keyaddr);
-            }
-        }
-
-        let (sink, stream) = Bytes.framed(stream).split();
-
-        let mles_rx = mles_rx.map_err(|_| panic!()); // errors not possible on rx XXX
-
-        let mles_rx = mles_rx.and_then(move |buf| {
-            if buf.is_empty() {
-                return Err(Error::new(ErrorKind::BrokenPipe, "broken pipe"));
-            }
-            if None == key {
-                //create hash for verification
-                let decoded_message = Msg::decode(buf.as_slice());
-                keys.push(decoded_message.get_uid().to_string());
-                keys.push(decoded_message.get_channel().to_string());
-                key = Some(MsgHdr::do_hash(&keys));
-                cid = Some(MsgHdr::select_cid(key.unwrap()));
-            }
-            let msghdr = MsgHdr::new(buf.len() as u32, cid.unwrap(), key.unwrap());
-            let mut msgv = msghdr.encode();
-            msgv.extend(buf);
-            Ok(msgv)
-        });
-
-        let send_wsrx = mles_rx.forward(sink);
-        let write_wstx = stream.for_each(move |buf| {
-            let ws_tx_inner = ws_tx.clone();
-            // send to websocket
-            let _ = ws_tx_inner.send(buf.to_vec()).wait().map_err(|err| {
-                Error::new(ErrorKind::Other, err)
-            });
-            Ok(())
-        });
-
-        send_wsrx
-            .map(|_| ())
-            .select(write_wstx.map(|_| ()))
-            .then(|_| Ok(()))
-    }).map_err(|_| {});
-
-    runtime.spawn(client);
-
-    match runtime.run() {
-        Ok(_) => {}
-        Err(err) => {
-            println!("Error: {}", err);
-        }
-    };
-}
-
